data_magar.py copied to *_old.py and new data manager started
